Specify the use of credentials in geojs tiles.
Chrome has a bug associated with the crossOrigin parameter of an image request.  If you set crossOrigin='anonymous', Chrome will not send cookies even for the a non-cross origin request.  If it is set to 'use-credentials', it works.  In the version of geojs referenced here, we always set the crossOrigin parameter to one of these two values.  In the future, geojs will be modified to NOT set the crossOrigin parameter when the tile request is on the same origin.