Added additional tile test options.
Confirmed that this works with non-square tiles (and tiles of any size).

The test tile server now can take a variety of query parameters:
  min: minimum tile level (default 0)
  max: maximum tile level (default levels parameter - 1)
  fractal: if 'true', draw a simple fractal on the tiles (only on square power-of-two tiles)
  w: tile width in pixels (default tileSize parameter)
  h: tile height in pixels (default tileSize parameter)
  sizex: image width in pixels at maximum level.
  sizey: image height in pixels at maximum level.
The tile size is percolated through to the viewer javascript as tileWidth and tileHeight, if available.  If additional viewers are extended to handle non-square tiles, they should check if these values are set and fall back to tileSize if not.

The following viewers work with non-square tiles:
  GeoJS
  OpenLayers
  OpenSeaDragon
Leaflet claims it supports non-square tiles by using L.Point for the tileSize, but the code which detects retina displays breaks when it encounters this, so it doesn't work as simply as the documentation claims.  I haven't tried to use non-square tiles in SlideAtlas.