Fix some issues in creating and deleting tile sets.
Instead of testing if a file is a tile file, check if it has pyramidal tiles.  If not, run it through the conversion process.

Only check that there are at least 3 channels rather than exactly 3.  The PHOTOMETRIC check ensures that the first three channels are what we expect.  This allows, for instance, uploading a PNG and using it despite having an unused opacity channel.

When we fail to find tiles in a file, include the last exception as part of the error message.  This makes it easier to determine why we couldn't handle the tiles (for instance, because it is in the wrong color space).