Throw an exception when instantiating the DummyTileSource.
When we installed the large_image plugin in an environment that had openslide-python but not the libtiff module, the DummyTileSource replaced the tiff tile source (as intended), but then this "processed" SVS files and prevented them from working.  This change will allow the file to be properly handed off to the SVS tile source.

Improve the error messages for import failures.

Added some tests for import failures.