Refactor how we load tile sources.
If a tile source fails to load, don't make it available for use (before we substituted the dummy source).  All sources now have a canRead class method which determines if they can read a file.  For test and dummy (more generally, for a tile source which inherits from TileSource and doesn't override it), these return false.  For tiff and svs (more generally, for a tile source which inherits from GirderTileSource and doesn't override it), these try to instantiate a tile source with the item and return if this can succeed.

The list of available tile sources is now generated by the tilesource module, and not separately by models/image_item.

Tile source import code has less checks, as the module does the checking in one spot.