Basic REST functionality is present.
Annotation elements are stored in a separate model (annotationelement).  These are created and deleted along with the annotation.  They are also completely deleted and recreated when an annotation is modified -- this causes database churn that could be skipped with an intelligent update method.  However, since the elements are ordered, maintaining the order could result in nearly as much churn as deletion and recreation.

There is another gotcha here: without database transactions, there will be a time when the annotation and the annotation elements are out of sync.  Furthermore, if we delete the elements before saving them (as currently), there will be a brief time when there are NO elements.  If we save before delete, there would be a time where there are BOTH sets of elements.

A way to work around this, wherein the new elements are always returned with the updated annotation, is to tag each element with a version that matches a version in the annotation.  The elements would be written BEFORE the annotation, and old elements deleted AFTER the annotation.  I'll implement this next.

There is a bit of monkey-patching of the basic model load/save/remove functions to ensure the event behavior with the annotations split across two models.