Add an ID to annotation elements.
The element ID defaults to the Mongo ID at the time of creation.  If the id is not passed back during a PUT modification, a new ID will be created for the element.  The element ID must be of the same form as a Mongo ID (24 hex digits).

Ensure atomicity in reading and writing annotations and their elements.  Because Mongo is not a transaction database, this is significantly complicated.  I am maintaining a monotonic sequence in the annotationelement collection that is used to ensure consistency.  I think I've covered the corner cases of multiply clients affecting the mongo database concurrently, but need to write tests to ensure that they are truly covered.