Add a getTileSource function.
You can now import large_image, which exposes tilesource (useful for exceptions) and a getTileSource function.  A python client can call getTileSource(<path or URI>[, user=<girder user>], **kwargs) to get any available tile source.  For girder items, this is of the form girder_item://(item id).  For the test and dummy tile sources, this is large_image://test and large_image://dummy.  For files, this is the path.  Eventually, we could support other protocols (http, for instances), where we handle it in a file-like manner for the file tile sources.