Add setup.py.
Rearranged the __init__.py and large_image.py code to not be in the repository's root.

Fixed an over-restrictive permissions issue getting a girder item from getTileSource.

Updated requirements.  Note that libtiff requires numpy, but doesn't list that as a requirement in its packaging, so you need to install numpy before you can install large_image.