Basic benchmarking to grab tiles from Girder
Takes an image id and attempts to request all tiles from all levels
of the image stored on Girder. Elapsed time for each request is recorded
and is used to generate a histogram and a box and whiskers plot.
Data is stored in a text file in the same directory as the script.