Ploting is now performed after generating files containing elapsed time data
Benchmark.py takes in a config file as an argument and requests images
from the servers listed in the config file. Elapsed times for each image is
stored in a file, there is one file per server being queried.
Plot_time.py takes a list of files and plots the elapsed time data generating
three histograms .