Add pixels and mm to units.  The previously named pixels units is now base_pixels.
Rename _getMagnification to getNativeMagnification.

Document CMake issue better (there is a patch submitted to the CMake mailing list to fix the bug, but until then commenting out a line in the plugin.cmake file is the solution).

Improve travis to build the plugin using more cmake flags and less ctest parameters.