Limit the number of simultaneous thumbnails that are requested in the itemList.
Chrome limits the maximum connectors to a domain to 6.  Girder uses one for getting events, leaving five.  This uses no more than 3 for thumbnails, which leaves some overhead for navigation.  If you navigate between item lists quickly, the connections can still be saturated, as Chrome doesn't cancel the image requests.  We would have to load the image to a hidden element first, and then load it to the visible element to work around that, so that the hidden elements don't vanish on page transitions.