Allow small images to be used as single tiles.
When used as a Girder plugin, the maximum size that is allowed is stored in a setting (a value of 0 disables direct use of small images).  When used without Girder, the maximum size is specified in server/tilesource/pil.py.  In either case, it can be overridden when asking for a tile source with the maxSize parameter.