Add edge options for tiles.
Tiles at the right and bottom edges of images can now be cropped or have unused areas converted all to a specific color.  With appropriate query strings, this allows all viewers to show the edges of images in correct manner.

Some of the tile output code has been moved into the base class to avoid repetition.  As a benefit, the TIFF tilesource can output tiles as PNGs if so desired.

Removed some vestigial code related to class cache keys.