Add functions for returning a single tile
This adds two functions: getSingleTile and getSingleTileAtAnotherScale. These functions take exactly the same parameters as tileIterator, but will return the first tile from the iterator.

tileIterator now takes an optional parameter tile_position which may be a 0-based integer which is functionally the same as iter(list(tileIterator(...))[tile_position]). tile_position may also be a dictionary of either {i, j} or {x, y}, where {i, j} are the 0-based tile positions within the iterator, and {x, y} are the level-based locations of the tile at the iterated level.

The tile information yielded by the iterator now includes tile_position and iterator_range which include information about the tile position in the iterator and the ranges of x, y, i, j, and position that the iterator produces.