The LRU cache was using the wrong key.
The metaclass cache was using some of the query parameters for the keys, resulting in few cache hits for certain functions (like getRegion).  This fixes the cache keys.

This still needs tests, and tile sources are not closed in a timely manner (even without caching), resulting in too many open files after a few hundred distinct sources.