Better tile source caching.
Prior to this, each tile source class used a separate cache.  This makes memory use difficult to manage intelligently.  The cache size needs to be set based on available file handles, as that is typically the limiting factor.  TIFF tile sources tend to have one open file handle per level of tiles, so the maximum number of tile sources that can be cached is now set based on that resource if it can be queried.  Moreover, ask to use more file handles if possible (it may not be).