Add an endpoint to clean up items with unfinished large_image jobs.
If a large image job fails to complete (for instance, because Girder is restarted during the process, or the worker is not running), the item is marked as having a large image in progress, which prevents easily restarting the job.  This endpoint cleans up these states, cancelling any large_image job that is in progress or stuck.