JFIF headers are now conditional.
See if an iOS is requesting a tile.  If so and we don't have a JFIF header, convert the JPEG to a PIL image and then to a new JPEG rather than adding a JFIF header.

It turns out that the colorspace of a JPEG is complicated, and adding a JFIF header changes it in unexpected ways.  See the comments.