Make an option for how edges are handled for overlapping tiles.
From the tileIterator function documentation of the tile_overlap['edges'] parameter:

As an example, suppose an image that is 8 pixels across (01234567) and a tile size of 5 is requested with an overlap of 2.  If the edges option is False (the default), the following tiles are returned: 01234, 12345, 23456, 34567.  Each tile reports its overlap, and the non-overlapped area of each tile is 012, 3, 4, 567.  If the edges option is True, the tiles returned are: 012, 0123, 01234, 12345, 23456, 34567, 4567, 567, with the non-overlapped area of each as 0, 1, 2, 3, 4, 5, 6, 7.

In either case, the non-overlapped area of all tiles is the same.