Use python as the default cache when girder isn't available.
By default, we had been using memcached.  When Girder isn't available, we are likely running in Dask or another environment where memcached is probably not configured properly.  As such, use the python caching backend as the default in those circumstances.  This can always be overridden.

Added a comment on the example showing how to specify the caching method explicitly.