Add client paging of annotations.
This first fetches the largest annotation elements, limited by a total detail level.  If this is the entire set of elements, they are just shown and no paging is used.  Otherwise, as the map is panned, zoomed, etc., the annotations are paged whenever the map has moved substantially (at least 1 zoom level or half the width/height of the viewport).  When zoomed out, the elements are limited by absolute size.  If more elements need to be shown than some arbitrary limit, only the largest elements are shown.