Guard against LRU cache issues.
cachetools LRU cache is not thread safe.  If two threads each cause an eviction, it would throw a KeyError and never recover.  Our cache would then fail, functionally breaking large_image until the server was restarted.

This does two things: (1) use a threading lock for the LRU cache, just as we do for memcached.  This should prevent the problem from ever occuring, and (2) guard against such a KeyError, and, if it occurs, return the cache miss value in any case.  If the LRU cache is used in a multithreaded environment without a lock, this guards against the catastrophic failure.

Tests have been added to ensure that multithreaded use works both with and without a cache lock.