Support content disposition on some endpoints.
On the region, associated image, and thumbnail endpoints, support requesting a content disposition.  The contentDisposition query parameter can be left blank for the existing behavior, or set to 'inline' or 'attachment'.  The filename is absed on the item name with the word 'region', 'thumbnail', or the associated image label appended to it and the appropriate suffix for the mimetype, except that cached thumbnails yield a generate '_largeImageThumbnail' filename.  The filename is encoded in a safe ASCII manner and a UTF-8 manner such that it should work on most clients.

This is based on the small-tif-output branch to leverage some of the testing infrastructure there.