Add a method to delete an annotation without unbinding listeners
The backbone API for destroying models also removes all event listeners.
To avoid regenerating an annotation model (and updating all views that
listen to it), this adds a new public that performs the rest request
but does not destroy the model object itself.  Very little code
is changed here; it is mostly an additional test for the new public
method.