Fix resample option for the tile iterator.
When the resample option was True (or one of the PIL.Image.NEAREST, LANCZOS, BILINEAR, or BICUBIC), there were some instances where resampling would fail due to a float/integer requirement mismatch.  Further, before, we would resample and scale AFTER the tile was computed and generated, which yielded tiles that were not the requested size.  For instance, asking for a tile of 1024 pixels square and a magnification that is 0.625 of a value of an existing level would result in tiles that were 640 square.  Now, there are fewer tiles but they are the requested size.  This also changes when no tile size is explicitly requested -- the tiles are now the size of the tiles at an native magnification.