Add a redirect parameter to GET tile.
If redirect is not specified, HTTP redirection will not be used.  If, however, the tile is stored as a file in the right format, it will be read and returned without reencoding.  If redirect is 'exact', redirection will only occur if the original tile is in the same encoding (JPEG, PNG, etc.) and quality settings; if 'encoding', then the result may have different quality settings; if 'any' then either a JPEG or a PNG may be returned.