Add annotation history.
This exposes a setting on the plugin settings page to toggle whether
history is saved.  It defaults to on.

If history is turned on, old versions of annotations are maintained in
the annotation and annotationelement mongo collections.  There is a
strictly monotonic version number for all annotations.  When an
annotation is updated, the old elements are retained, and the old
version of the annotation is saved with an _active=false marker.  These
old versions have an _annotationId to reference the original annotation.

Historic versions can be listed via a new endpoint.  Historic versions
can be fetched via the original annotation id and version number or via
the historic record _id (all with appropriate user level
authentication).  A revert endpoint allows switching to previous
versions of an annotation.