Refactor to allow different projections.
The mapnik source now takes an option projection parameter.  If not
specified, the source is in pixel coordinates.  If specified, it is in
the specified proj4 projection.

The default thumbnail in girder will appear in pixel space coordinates.

Move the mapnik source test to its own test file.

Don't test mapnik source in travis with Python 3.

Also, update mongo versions in CI.