Add a getPixel method and rest endpoint.
Internally, `getRegion` can now specify width and height in different
units than the edges of a region.  This allows a 1-pixel square region
to be specified with the left and top specified via non-pixel units.
`getPixel` gets a 1x1 pixel region and returns a tuple of the channel
values as an object.  Subclasses of the tile source can augment this
with additional information (such as band data).

For the mapnik source, this also returns band information at the pixel.

Also, the generated hash key for methods in cached classes now includes
the class hash key.