Allow filtering by image name in annotated image search
This commit moves most of the logic from the /annotation/images endpoint
from the rest endpoint into a model method and adds a new keyword
parameter for filtering by image name.  As before, this filtering is
done purely in python, but we may want to refactor this method to use an
aggregation pipeline if performance becomes a concern.