Allow girder to show load errors in more cases.
If girder is installed but requirements of large_image are not, the
plugin would report as enabled but not be functional.  This checks if
girder is available, and, if so, allows girder to report errors.

Eventually, it will be nice to separate large_image without girder from
the necessary girder parts to make this easier to manage.  For now, this
should make diagnosis easier.

As a simple test, when everything works under girder, uninstall a
required module (cachetools, for instance), and restart girder.  Before,
no error was reported.  With this change, the missing module is
reported.