Improve level power-of-two detection.
For SVS tile sources, we check if each level is a power of two of the
base level.  Using `abs(modf(log2ratio)[0])` to get the mantissa, this
could sometimes return a failure (such as with 45808 and 11456), since
the mantissa was near to 1 rather than near to zero.  Using `round` to
get the mantissa is better.