Change the order a new annotation and its element are stored.
Before, we wrote (inserted) an `annotation` to the mongo collection,
then wrote the set of `annotationelement`s.  If an item's annotations
were listed before this finished, the new annotation would be included
in the list.  Requesting the annotation could return partial results for
the annotation elements.  By saving the annotation after the annotation
elements, this problem is avoided.