Change sorting of Annotation collections
The default girder behavior is to sort by name, which is not an
attribute used by the annotation model.  The default sorting behavior
has been changed to sort by most recently created.  This behavior is
explicitly disabled in the annotation list widget to preserve the order
returned by the server.