Fix an issue when drawing selection rectangles.
If, for instance, in HistomicsTK, you select a region by clicking and
dragging, it works fine.  But if you select the same region by clicking
opposite corners of the rectangle, the rectangle is deformed.
Specifically, this computes center, width, and height of a rectangle,
and made two assumptions: that the first point was always a specific
corner, and that the points always were in the same winding order.
Neither of these assumptions are always correct.  This avoids making
assumptions, but otherwise maintains the same process.