Change a few cache defaults.
If config is not set, then we aren't running in Girder, in which case
use a smaller memory portion for caches.  A common use case is Dask with
multiple workers, and the ordinary default would ideally be divided by
the maximum number of workers on any given machine.

Also, double the expected memory per image tile by default.