Fix clobbering annotation elements when changing access.
We avoid loading annotation elements into memory when we don't need to
work with them, but we were saving the entire annotation document when
changing access permissions.  This functionally deleted the elements.
Rather, update the annotation record to update the acl.  This is both
more efficient and avoids the problem.

This fixes https://github.com/DigitalSlideArchive/HistomicsTK/issues/537