Speed up annotation output.
We stream annotation elements since outputting them as a single blob of
json can use a lot of memory and causes a delay between fetching them
from the database and emitting them.  However, emitting elements one at
a time is slower then emitting them in batches.  This batches 100
elements at a time, which, experimentally, provides a balance between
good speed and rapid start.

As an example, a whole slide image with ~370,000 detected nuclei
annotations when from taking ~9 seconds to emit the top 25000
annotations to taking ~2.4 seconds.