More robust access optimization for tile download route
The history here is that we deploy large_image into some environments
in which the access.public decorator has its behavior changed to be
equivalent to access.token, and the optimization used to facilitate
cookie-based access to the tile endpoint does not work in that condition.
By also removing access.public and explicitly disabling the access
level checking, we can make this optimization work in that environment.