Change the tileIterator resample default to True.
Prior to this, resample defaulted to False.  This led to surprising
results when a magnification was asked for that was not a native tile
level in a source image, such as svs files where not all levels are
present.

When tiles need to be resampled, make the tile['magnification'] value
reflect the output magnification and the the tile['tile_magnification']
value be the magnification before resampling.

Improve some documentation for the resample parameter.