Add the ability to copy annotations with items.
This adjusts the POST resource/copy, POST item/{id}/copy, and POST
folder/{id}/copy endpoints.  Each of these endpoints take an additional
boolean parameter of 'copyAnnotations'.

Note that the item is copied and then the annotations are copied, so, if
the annotations are large, you will see the item in the browser and the
annotations will show up a brief time later.  The endpoint only returns
when the entire copy is done.

Copied annotations will have the same permissions as the copied item.