Use a consistent hash function for long keys and memcached.
We had been using Python's hash() function to reduce a long key to a
shorter key for the cache keys in memcached.  However, Python 2 can be
optionally set to use a random or non-standard seed for hash(), and
Python 3 uses a random seed for hash() by default.  This means that
subsequent runs of a program would not utilize the memcached tile cache
from previous runs.

Also, improve the default class key for the mapnik tile source, as this
allows easier programmatic reuse of tiles.