This improves handling tiff files with multiple tiled images.
Before, if there are multiple tiled images (each possibly with a pyramid
of smaller images), the image with the largest tiles was selected.  Now,
the image with the largest tiles that are not the whole image are
selected.

Also, this adds parsing for Optrascan metadata stored as xml within the
pixel space of an uncompressed image.