Fix an issue with annotation permissions.
If a user or group has been removed from the system such that it still
exists in an access control list, when the access list is read, Girder
prunes the missing users and groups and writes back the access list.
However, for annotations, this could cause a save with an empty array of
elements, effectively removing the elements from the annotation.  This,
instead, updates the annotation's access list.