Do a migration of database values on start.
We were doing on-demand migrations to have ACLs and groups in annotation
models, since originally we didn't have these.  Not having them prevents
certain optimizations, like getting count of total number of annotations
as a header when listing annotations.  Although there could be a slow
one-time response to this, and, without appropriate indices, the initial
start-up will be slower, the benefits outweigh this cost.

As a benefit of this, for instance, the query to display the number of
annotations in the image list can return much less data, since it can
get the number of annotations from the header.