Speed up item list annotation counts.
When we show an item list in Girder with annotations, we show a badge on
each thumbnail indicating how many annotations are on that large image.
Before, this made one REST request per call without any sort of
throttling, which could effectively block navigation until enough of the
calls were resolved.  This gets all needed values in a single REST
request.