Work around a file descriptor issue in cheroot.
There is an issue that was introduced in 8.1.0 that throws an error when
file descriptor numbers are large (see
https://github.com/cherrypy/cheroot/issues/249).  We can't pin cheroot
to an older version, as that breaks support for modern python versions.

It isn't clear if limiting the size of our cache will really help -- the
cheroot issue implies that it is leaking file descriptors, so this may
not do anything useful.