Refactor image conversion task.
This splits the conversion process so that it is separate from the
Girder worker task.  The new packages are now in the utilities
directory.

It adds explicit support for processing geotiffs.  It adds support for
reprocessing files that could be read by large image, making them a more
preferred format for reading tiles.

There are now options for lossless or lossy compression, plus more
options for compression techniques.  By default, if a file is not known
to be compressed in a lossy manner, the resulting converted file will be
lossless.

The converter doesn't use subprocess (it uses the python interfaces for
vips and GDAL); this keeps the memory footprint smaller.

This is part of a series of PRs refactoring conversion.  Future PRs will
add support for multi-frame images, for converting files that can be
read with large image but not by vips or gdal, and for adding jp2k tile
support.