Harden updating the item view after making a large image.
This is one of the causes of transient test failures, but shows a user
interface bug.  Occasionally, when making a large image that requires
converting via a job, the page would update after the job completed but
before the image was marked as updated.  This would result in the page
refreshing and still showing that the job was pending.

This substantially reduces transient test failures in one test
environment.