Fix an issue with non-integer GDAL sources.
If a GDAL source had non-integer data, the band maxim/min calculation
would fail.  Specifically, we were adding the maximum range of the
integer type as a value at the end of the list of data for that frame's
bands.  If the dtype is not an integer type, this fails.