Reduce open files in conversion.
Before, reading tiles was parallelized, then converting the composited
vips image was separately parallelized.  vips ends up keeping a file
handle open per tile until the image is closed.  For slow conversion
processes (such as jp2k or webp), this could result in the majority of
the frames being ready to process all waiting on the first few frames to
be written, with a file handle open per tile per frame.  This changes
the size of the tiles to reduce the count and parallelizes at the frame
level rather than separately at the tile and the save level (tiles are
still read in parallel).

Note that for systems with a huge number of processors, we could still
have or super large images, we could still exhaust the file handles.
Probably the real solution to this would be to dynamically pick a tile
size such that the number of tiles would be under some threshold,
trading memory use for file handle use, and, likewise, having some cap
to the number of frames that are processes in parallel.  Another
possibility would be to raise the per-process file descriptor limit.