Provide an option to output large regions as tiled tiffs.
This also will geotag geospatial outputs.

Specifically, when calling the getRegion function or endpoint, if an
encoding of `TILED` is used (previously this could have been one of
`JPEG`, `PNG`, or `TIFF`), a tiled tiff is output.  For non-geospatial
sources, this is tiled using vips.  For geospatial sources, this is
tiled using gdal.

When using the `TILED` option, a variety of parameters that can be used
in conversions can be used: `compression`, `quality`.  For
compatibility, these can also be `tiffCompression` and `jpegQuality`.