Add more cache configuration options.
Added cache_tilesource_memory_portion and cache_tilesource_maximum
config options.  The number of tile sources that are caches is the
smallest of (a) the number of available file handles divided by a factor
and reduced by a constant, (b) the available memory divided by the
portion (default 8) and the expected size of the largest tile source,
(c) the cache maximum value.  Before, the portion was always 8 and there
was no way to specify a maximum value.

This can be applied in girder, for instance via the [large_image]
section of the config file with something like::

    cache_tilesource_memory_portion = 16
    cache_tilesource_maximum = 64