Add a deepzoom tile source.
This works with the dzi files as output by vips in both onetile and
onepixel modes and with both png and jpeg compression.  It should allow
for any tile size and overlap as generated by vips.

A future expansion would be to read property files and handle center and
other options in dzsave.

Stop paralellizing tests that use the girderWorker fixture, as the
session scope doesn't do what was expected in pytest-xdist.