Improve tile source locking.
Before, if two different tile sources using the same cache were being
opened simultaneously, they would be opened sequentially due to the
locking semantics.  This also prevented opening the same source twice.

Now, when a tile source is not in cache, a separate per-key lock is
created for opening that source.  This prevents the same source from
being opened twice but allows different sources to be opened
concurrently.