Add property to control clamping bounds
New property `_unclampBoundsForOverlay` with getter/setter added to control
whether or not the X/Y bounds should be clamped when rendering image overlays.
Unclamping the bounds is useful because image overlays may be cut off by the viewer
if they are larger than the base image layer. However, unclamped bounds eliminates
the scroll out to center functionality. The default behavior at the time of this
commit is to unclamp the bounds.