The tile iterator could return excess tiles with overlap.
This fixes a bug where, when overlap was specified, some smaller tiles
could be returned at the right/bottom edges that were not necessary.
Since the overlap was documented, this probably only resulted in
inefficiency, not and breakage.  This change now conforms to the
tileIterator documentation and adds a test to ensure this is the case.