Add an option to including bounding box information with elements.
This adds an option to the GET annotation/{id} endpoint and to the
annotationelement.yieldElements method.  The
annotationelement.getElements and annotation.load methods support
passing the additional bbox=True argument as part of the region request.

The result that includes the _elementQuery record also has the bounding
box for the specific query (it is of just the elements that are returned
if they are paged or limited).

Closes #850.