Use functions in styles.
Resolves #688.

This can be used, for instance, to get a mask from a superpixel image.
For instance, suppose you want a mask for the superpixel with an index
of 82, you'd use a style like:
{"function":{"name":"large_image.tilesource.utilities.maskPixelValues","context":true,"parameters":{"values":[82]}},"bands":[]}
(or, if you are using superpixels with boundaries, this would be
{"function":{"name":"large_image.tilesource.utilities.maskPixelValues","context":true,"parameters":{"values":[164,165]}},"bands":[]}).
The resulting image (assuming you are returning a PNG), will be opaque
black where this value is not present and transparent white where it is
(or specify the negative and positive parameters to get different
colors).