Add a dicom source.
This uses wsidicom to read dicom files.  So far it has only been tested
with files that were generated with the wsi2dcm docker from the 1.0.3
deb from https://github.com/GoogleCloudPlatform/wsi-to-dicom-converter,
which seems to have issues with some files.

This may be refactored to use a different base library, and internal
details should be considered provisional.

It needs to have internal metadata exposed.